+++ pokeranking/server/Models/Criatura.js
